// src/contracts/generated/maverick-schema.d.ts
export interface components {
  schemas: {
// ...
    Event: {
      /** Format: int64 */
      doc?: number;
      /** Format: int64 */
      fields?: number;
      /** Format: int64 */
      id: number;
      name: string;
      /** Format: int64 */
      venueId: number;
      /** Format: int64 */
      venueConfigurationId: number;
      /** Format: date-time */
      eventTimeLocal: string;
      /** Format: date-time */
      endDate?: string;
      timeTbd: boolean;
      dateTbd: boolean;
      url?: string;
      parking: boolean;
      /** Format: int64 */
      parkingEventId?: number;
      /** Format: int32 */
      minimumAge?: number;
      ifNecessary: boolean;
      multiDate: boolean;
      /** Format: date-time */
      eventTimeUtc?: string;
      /** Format: int64 */
      cmsPageId?: number;
      eventPerformers?: components['schemas']['EventPerformer'][];
      venue?: components['schemas']['Venue'];
      slug?: string;
      venueConfiguration?: components['schemas']['VenueConfiguration'];
      seasonTicket: boolean;
      /** Format: date-time */
      nextOnsaleTimeLocal?: string;
      /** Format: date-time */
      nextPresaleTimeLocal?: string;
      /** Format: date-time */
      lastOnsaleTimeLocal?: string;
      /** Format: date-time */
      lastPresaleTimeLocal?: string;
      /** Format: date-time */
      firstOnSaleTimeLocal?: string;
      /** Format: int32 */
      daysToEvent?: number;
      /** Format: int32 */
      daysToOnsale?: number;
      /** Format: int32 */
      daysToPresale?: number;
      /** Format: int32 */
      daysUntilFirstOnSale?: number;
      /** @enum {string} */
      pricingType?: 'DEFAULT' | 'ALL_IN_TICKET_DETAILS' | 'ALL_IN_TICKET_PAGE';
      /** Format: int32 */
      availableTickets?: number;
      /** Format: double */
      percentInventoryAvailable?: number;
    };
    EventPerformer: {
      /** Format: int64 */
      id: number;
      /** Format: int64 */
      performerId: number;
      master: boolean;
      performer?: components['schemas']['Performer'];
    };
    GiftCardRedemption: {
      amountRedeemed: number;
      /** Format: int64 */
      giftCardId: number;
    };
    Listing: {
      /** Format: int64 */
      id: number;
      /** Format: int64 */
      eventId: number;
      section: string;
      /** Format: int64 */
      sectionId: number;
      row: string;
      notes?: string;
      /** Format: int32 */
      quantity: number;
      validSplitQuantities: number[];
      flex: boolean;
      instant: boolean;
      displayPrice: number;
      serviceFee?: number;
      powerSeller: boolean;
      /** Format: date */
      inHandDate?: string;
      stockType: string;
      deliveryMethods: components['schemas']['DeliveryMethod'][];
      attributes?: number[];
      seatNumbers?: string;
      generalAdmission: boolean;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}